renamed knowledgegraph to knowledge_graph.py
no problemo @mads @kaare @kasper nielsens
